moved providers to separate file
providers have been moved to a separate file and condensed so they aren't all on App.tsx
