+++ client/src/context/index.tsx
@@ -1,0 +1,36 @@
+import {
+  ApolloClient,
+  ApolloProvider,
+  createHttpLink,
+  InMemoryCache,
+} from '@apollo/client';
+import { BrowserRouter } from 'react-router-dom';
+import { setContext } from '@apollo/client/link/context';
+
+const httpLink = createHttpLink({
+  uri: '/graphql',
+});
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem('id_token');
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
+const client = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+interface ProviderProps {
+  children: React.ReactNode;
+}
+
+export const Providers = ({ children }: ProviderProps) => {
+  return (
+    <ApolloProvider client={client}>
+      <BrowserRouter>{children}</BrowserRouter>
+    </ApolloProvider>
+  );
+};
